Switch to using client_credentials strategy.
This strategy allows you to authenticate with just the client_id and
client_password. However, it normally limits access to a basic level -
nothing that is per-user.

The Zoning API has been configured to allow full access to 'trusted'
applications without authentication. This means that for trusted
applications, the client_credentials strategy gives the same level of
access as the other strategies.